+++ api/src/infra/database/database.module.ts
@@ -1,8 +1,8 @@
 import { Global, Module } from '@nestjs/common';
 import { PrismaService } from './prisma.service';
-import { UsersRepository } from './repositories/users.repositories';
-import { CategoriesRepository } from './repositories/categories.repositories';
-import { BankAccountsRepository } from './repositories/bank-accounts.repositories';
+import { UsersRepository } from './repositories/users.repository';
+import { CategoriesRepository } from './repositories/categories.repository';
+import { BankAccountsRepository } from './repositories/bank-accounts.repository';
 
 @Global()
 @Module({
