+++ api/src/modules/auth/auth.service.ts
@@ -3,7 +3,7 @@ import {
   Injectable,
   UnauthorizedException,
 } from '@nestjs/common';
-import { UsersRepository } from 'src/infra/database/repositories/users.repositories';
+import { UsersRepository } from 'src/infra/database/repositories/users.repository';
 import { compare, hash } from 'bcrypt';
 import { JwtService } from '@nestjs/jwt';
 import { SignupDto } from './dto/signup.dto';
